requirements.txt
pandas 
python-docx 
pyyaml
openpyxl